io/audio: replace bzero/bcopy; fix size_t length math in audio paths (reduce implicit widening, no behavior change)

+++ src/core/audio/dsd_audio.c
@@ -652,17 +652,17 @@ playSynthesizedVoice(dsd_opts* opts, dsd_state* state) {
         if (opts->audio_out == 1 && (opts->audio_out_type == 5 || opts->audio_out_type == 1)) //OSS
         {
             //OSS 48k/1
-            result =
-                write(opts->audio_out_fd, (state->audio_out_buf_p - state->audio_out_idx), (state->audio_out_idx * 2));
+            result = write(opts->audio_out_fd, (state->audio_out_buf_p - state->audio_out_idx),
+                           (size_t)state->audio_out_idx * 2u);
             state->audio_out_idx = 0;
         } else if (opts->audio_out == 1 && opts->audio_out_type == 0) {
             pa_simple_write(opts->pulse_digi_dev_out, (state->audio_out_buf_p - state->audio_out_idx),
-                            (state->audio_out_idx * 2), NULL);
+                            (size_t)state->audio_out_idx * 2u, NULL);
             state->audio_out_idx = 0;
         } else if (opts->audio_out == 1
                    && opts->audio_out_type == 8) //UDP Audio Out -- Forgot some things still use this for now
         {
-            udp_socket_blaster(opts, state, (state->audio_out_idx * 2),
+            udp_socket_blaster(opts, state, (size_t)state->audio_out_idx * 2u,
                                (state->audio_out_buf_p - state->audio_out_idx));
             state->audio_out_idx = 0;
         } else {
@@ -692,18 +692,18 @@ playSynthesizedVoiceR(dsd_opts* opts, dsd_state* state) {
         {
             //OSS 48k/1
             result = write(opts->audio_out_fd, (state->audio_out_buf_pR - state->audio_out_idxR),
-                           (state->audio_out_idxR * 2));
+                           (size_t)state->audio_out_idxR * 2u);
             state->audio_out_idxR = 0;
         } else if (opts->audio_out == 1 && opts->audio_out_type == 0) {
             pa_simple_write(opts->pulse_digi_dev_outR, (state->audio_out_buf_pR - state->audio_out_idxR),
-                            (state->audio_out_idxR * 2), NULL);
+                            (size_t)state->audio_out_idxR * 2u, NULL);
             state->audio_out_idxR = 0;
         } else if (
             opts->audio_out == 1
             && opts->audio_out_type
                    == 8) //UDP Audio Out -- Not sure how this would handle, but R never gets called anymore, so just here for symmetry
         {
-            udp_socket_blaster(opts, state, (state->audio_out_idxR * 2),
+            udp_socket_blaster(opts, state, (size_t)state->audio_out_idxR * 2u,
                                (state->audio_out_buf_pR - state->audio_out_idxR));
             state->audio_out_idxR = 0;
         } else {
